improve my account page design
