+++ pwa/components/profile/ChangePassword.tsx
@@ -1,7 +1,7 @@
 import React, {useState, ChangeEvent, useEffect} from 'react';
-import {useAccount, useAuth} from '@contexts/AuthContext';
+import {useAccount} from '@contexts/AuthContext';
+import {authenticationResource} from '@resources/authenticationResource';
 import {userResource} from '@resources/userResource';
-import {validatePassword} from '@utils/passwordValidator';
 import {
   Box,
   Typography,
@@ -9,8 +9,9 @@ import {
   Button,
   CircularProgress,
   Alert,
+  Grid,
 } from '@mui/material';
-import {authenticationResource} from '@resources/authenticationResource';
+import {validatePassword} from '@utils/passwordValidator';
 
 const UpdatePassword = (): JSX.Element => {
   const {user} = useAccount({redirectIfNotFound: '/login'});
@@ -28,7 +29,6 @@ const UpdatePassword = (): JSX.Element => {
   const [checkPasswordErrorText, setCheckPasswordErrorText] = useState<
     string | null
   >(null);
-  const {login} = useAuth();
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
@@ -37,21 +37,20 @@ const UpdatePassword = (): JSX.Element => {
     }
     setErrorMessage(null);
     setPendingLogin(true);
-
     const response = await authenticationResource.checkCurrentPassword({
       email: email,
       password: oldPassword,
     });
-
     if (response.ok) {
       setSuccessOldPassword(true);
     } else {
-      setErrorMessage("Ce mot de passe n'est pas valide");
+      setErrorMessage("Ce mot de passe n'est pas valide.");
     }
-
     setPendingLogin(false);
+    setTimeout(() => {
+      setErrorMessage(null);
+    }, 3000);
   };
-
   useEffect(() => {
     if (user) {
       setEmail(user.email);
@@ -99,16 +98,18 @@ const UpdatePassword = (): JSX.Element => {
       setPendingLogin(false);
       return;
     }
-    if (checkPassword == newPassword) {
+    if (checkPassword === newPassword) {
       try {
         await userResource.putById(user.id, {
           email: user.email,
           plainPassword: newPassword,
         });
         setSuccessNewPassword(true);
+        setOldPassword('');
         setTimeout(() => {
           setSuccessNewPassword(false);
-          setOldPassword('');
+          setCheckPassword('');
+          setNewPassword('');
         }, 3000);
         setSuccessOldPassword(false);
       } catch (e: any) {
@@ -123,128 +124,139 @@ const UpdatePassword = (): JSX.Element => {
 
   return (
     <Box
-      mt={8}
-      p={4}
-      sx={{backgroundColor: 'grey.200'}}
-      width="100%"
-      borderRadius={2}
-      maxWidth="lg"
-      mx="auto">
-      <Typography fontSize={22} fontWeight={600}>
-        Modifier mot de passe
-      </Typography>
-      <Box
-        width="100%"
-        display="flex"
-        gap={{md: 4}}
-        flexDirection={{xs: 'column', md: 'row'}}
-        justifyContent="flex-end"
-        component="form"
-        onSubmit={handleSubmit}
-        noValidate>
-        <Box
-          display="flex"
-          flexDirection="column"
-          width={{xs: '100%', md: '40%'}}>
-          <TextField
-            sx={{backgroundColor: 'white', borderRadius: '20px', width: '100%'}}
-            margin="normal"
-            required
-            fullWidth
-            name="password"
-            label="Mot de passe actuel"
-            type="password"
-            id="password"
-            autoComplete="current-password"
-            value={oldPassword}
-            onChange={handleChangeOldPassword}
-          />
-          {!successOldPassword && (
-            <Box display="flex" flexDirection="column" alignItems="start">
-              <Button
-                type="submit"
-                variant="contained"
+      sx={{
+        mt: 1,
+        bgcolor: 'white',
+        px: {xs: 3, md: 5},
+        py: {xs: 4, md: 5},
+        boxShadow: 2,
+        width: {xs: '90%', md: '55%'},
+        borderRadius: 6,
+        mx: 'auto',
+        maxWidth: '700px',
+        position: 'relative',
+      }}>
+      <Box onSubmit={handleSubmit} component="form" noValidate>
+        <Grid container direction="column">
+          <Typography textAlign="center" pb={2} variant="h4" color="secondary">
+            Modifier mot de passe
+          </Typography>
+          <Grid container item xs={12} spacing={{xs: 0, md: 2}} direction="row">
+            <Grid item xs={12} sm={6} md={12} lg={6}>
+              <TextField
                 sx={{
-                  mt: 3,
-                  textTransform: 'capitalize',
-                  width: {xs: '100%', md: 'auto'},
-                }}>
-                {!pendingLogin ? (
-                  'Confirmer'
-                ) : (
-                  <CircularProgress size={20} sx={{color: 'white'}} />
-                )}
-              </Button>
-            </Box>
-          )}
-        </Box>
-        <Box
-          sx={{backgroundColor: 'white', width: '40%', visibility: 'hidden'}}
-        />
+                  backgroundColor: 'white',
+                  borderRadius: '20px',
+                  width: '100%',
+                }}
+                margin="normal"
+                required
+                fullWidth
+                name="password"
+                label="Mot de passe actuel"
+                type="password"
+                id="password"
+                autoComplete="current-password"
+                value={oldPassword}
+                onChange={handleChangeOldPassword}
+              />
+            </Grid>
+            {!successOldPassword && (
+              <Grid item xs={12} sm={6} md={12} lg={6}>
+                <Box
+                  display="flex"
+                  flexDirection="column"
+                  alignItems={{xs: 'center', md: 'end'}}>
+                  <Button
+                    disabled={!oldPassword}
+                    size="large"
+                    type="submit"
+                    variant="contained"
+                    sx={{
+                      mt: 3,
+                      textTransform: 'capitalize',
+                    }}>
+                    {!pendingLogin ? (
+                      'Confirmer'
+                    ) : (
+                      <CircularProgress size={20} sx={{color: 'white'}} />
+                    )}
+                  </Button>
+                </Box>
+              </Grid>
+            )}
+          </Grid>
+        </Grid>
       </Box>
-
       {successOldPassword && (
         <Box component="form" onSubmit={handleUpdatePassword} noValidate>
-          <Box
-            display="flex"
-            flexDirection={{xs: 'column', md: 'row'}}
-            justifyContent="flex-end"
-            gap={{md: 4}}
-            mt={1}>
-            <TextField
-              margin="normal"
-              required
-              fullWidth
-              name="newPassword"
-              label="Nouveau mot de passe"
-              type="password"
-              id="newPassword"
-              autoFocus
-              value={newPassword}
-              error={!!passwordErrorText}
-              helperText={passwordErrorText}
-              onChange={handleChangeNewPassword}
-              sx={{
-                backgroundColor: 'white',
-                borderRadius: '20px',
-                width: {xs: '100%', md: '40%'},
-              }}
-            />
-            <TextField
-              margin="normal"
-              required
-              fullWidth
-              name="checkPassword"
-              label="Confirmer le mot de passe"
-              type="password"
-              id="checkPassword"
-              value={checkPassword}
-              error={!!checkPasswordErrorText}
-              helperText={checkPasswordErrorText}
-              onChange={handleChangeCheckPassword}
-              sx={{
-                backgroundColor: 'white',
-                borderRadius: '20px',
-                width: {xs: '100%', md: '40%'},
-              }}
-            />
-          </Box>
-          <Box textAlign="right">
-            <Button
-              type="submit"
-              variant="contained"
-              sx={{
-                mt: 3,
-                textTransform: 'none',
-                width: {xs: '100%', md: 'auto'},
-              }}>
-              {!pendingLogin ? (
-                'Mettre à jour'
-              ) : (
-                <CircularProgress size={20} sx={{color: 'white'}} />
-              )}
-            </Button>
-          </Box>
+          <Grid container direction="column">
+            <Grid
+              container
+              item
+              xs={12}
+              spacing={{xs: 0, md: 2}}
+              direction="row">
+              <Grid item xs={12} sm={6} md={12} lg={6}>
+                <TextField
+                  margin="normal"
+                  required
+                  fullWidth
+                  name="newPassword"
+                  label="Nouveau mot de passe"
+                  type="password"
+                  id="newPassword"
+                  autoFocus
+                  value={newPassword}
+                  error={!!passwordErrorText}
+                  helperText={passwordErrorText}
+                  onChange={handleChangeNewPassword}
+                  sx={{
+                    backgroundColor: 'white',
+                    borderRadius: '20px',
+                  }}
+                />
+              </Grid>
+              <Grid item xs={12} sm={6} md={12} lg={6}>
+                <TextField
+                  margin="normal"
+                  required
+                  fullWidth
+                  name="checkPassword"
+                  label="Confirmer le mot de passe"
+                  type="password"
+                  id="checkPassword"
+                  value={checkPassword}
+                  error={!!checkPasswordErrorText}
+                  helperText={checkPasswordErrorText}
+                  onChange={handleChangeCheckPassword}
+                  sx={{
+                    backgroundColor: 'white',
+                    borderRadius: '20px',
+                  }}
+                />
+              </Grid>
+            </Grid>
+            <Grid item xs={12} sm={6} md={12} lg={6}>
+              <Box display="flex" flexDirection="column" alignItems="center">
+                <Button
+                  size="large"
+                  type="submit"
+                  variant="contained"
+                  sx={{
+                    mx: 'auto',
+                    mt: 3,
+                    textTransform: 'none',
+                  }}>
+                  {!pendingLogin ? (
+                    'Mettre à jour'
+                  ) : (
+                    <CircularProgress size={20} sx={{color: 'white'}} />
+                  )}
+                </Button>
+              </Box>
+            </Grid>
+          </Grid>
         </Box>
       )}
       {errorMessage && (
