+++ pwa/pages/mon-compte.tsx
@@ -1,10 +1,19 @@
 import {NextPageWithLayout} from 'pages/_app';
 import React from 'react';
 import Head from 'next/head';
-import {Typography, Box, CircularProgress, Container} from '@mui/material';
+import {
+  Typography,
+  Box,
+  CircularProgress,
+  Container,
+  Alert,
+  Button,
+  Grid,
+  TextField,
+} from '@mui/material';
 import {useAccount} from '@contexts/AuthContext';
 import WebsiteLayout from '@components/layout/WebsiteLayout';
-import MyAccountForm from '@components/profile/MyAccountForm';
+import MyInformations from '@components/profile/MyInformations';
 import RemoveAccount from '@components/profile/RemoveAccount';
 import ChangePassword from '@components/profile/ChangePassword';
 
@@ -17,26 +26,33 @@ const MyProfile: NextPageWithLayout = () => {
         <title>Mon compte</title>
       </Head>
       <WebsiteLayout>
-        <Container sx={{width: {xs: '100%', md: '80%'}}}>
-          {isLoadingFetchUser && (
-            <Box display="flex" justifyContent="center" my={10}>
-              <CircularProgress />
-            </Box>
-          )}
+        <Box
+          bgcolor="lightprimary.light"
+          height="100%"
+          width="100%"
+          position="absolute"
+          top="0"
+          left="0"
+          zIndex="-1"
+        />
+        <Container>
+          <Typography
+            textAlign="center"
+            variant="h1"
+            sx={{mt: 4}}
+            color="primary">
+            Mon compte
+          </Typography>
+
           {user && (
             <Box
-              width="100%"
+              py={4}
               display="flex"
               flexDirection="column"
-              alignItems="center">
-              <Typography
-                pt={4}
-                pb={6}
-                fontSize={{xs: 28, md: 30}}
-                fontWeight={600}>
-                Mon Compte
-              </Typography>
-              <MyAccountForm userLogged={user} />
+              gap={4}
+              position="relative"
+              alignItems="flex-start">
+              <MyInformations userLogged={user} />
               <ChangePassword />
               <RemoveAccount />
             </Box>
